Fix #4371. Reunite ConfigImportCommands classes now that they no longer need the UpdateKernel
